moved schain as a module
